+++ lib/kaze_client/request/requests/upload_attachment_request.rb
@@ -12,7 +12,6 @@ module KazeClient
   #   response = KazeClient::Client.new(URL).execute(rq)
   #   rq.send_attachment(response['direct_upload'])
   class UploadAttachmentRequest < Utils::FinalRequest
-
     include Utils::AuthentifiedRequest
 
     # @param attachment [ActiveStorage::Attachment]
@@ -24,9 +23,9 @@ module KazeClient
 
       @body = {
         blob: {
-          filename:     @attachment.filename,
-          byte_size:    @attachment.byte_size,
-          checksum:     @attachment.checksum,
+          filename: @attachment.filename,
+          byte_size: @attachment.byte_size,
+          checksum: @attachment.checksum,
           content_type: @attachment.content_type
         }
       }
@@ -36,7 +35,5 @@ module KazeClient
       header = direct_uploads['headers']
       HTTParty.put(direct_uploads['url'], { body: @attachment.download, headers: header })
     end
-
   end
-
 end