+++ CHANGELOG.md
@@ -4,6 +4,12 @@
 
 - Add new requests:
   - Upload activestorage attachment
+  - List the collections for the company of the current user
+  - Fetch the details for the given collection id
+  - List the items assigned to the collection id
+  - Fetch the details for the given item id from the given collection id
+  - Add tags to a job
+- Fix handling of 404 HTTP status.
 
 ## [0.3.1] - 2022-11-30